# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3333",
    "basePath": "/v1",
    "paths": {
        "/countries": {
            "get": {
                "description": "Endpoint to retrieve all countries",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get all countries",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of countries per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CountriesWithPagination"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Endpoint to create a new country",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Process a new country",
                "parameters": [
                    {
                        "description": "Country request payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CountryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CountryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/countries/{id}": {
            "get": {
                "description": "Endpoint to retrieve a country by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get country by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Country ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CountryResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CountriesWithPagination": {
            "type": "object",
            "properties": {
                "countries": {
                    "description": "Countries is the list of countries.\n\nrequired: true",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CountryResponse"
                    }
                },
                "limit": {
                    "description": "Limit is the number of items per page.\n\nrequired: true",
                    "type": "integer"
                },
                "page": {
                    "description": "Page is the page number.\n\nrequired: true",
                    "type": "integer"
                }
            }
        },
        "dto.CountryRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "description": "Name of the country\n\nrequired: true",
                    "type": "string"
                }
            }
        },
        "dto.CountryResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "The ID of the country\n\nexample: \"123e4567-e89b-12d3-a456-426614174000\"",
                    "type": "string"
                },
                "name": {
                    "description": "The name of the country\n\nexample: \"Brazil\"",
                    "type": "string"
                }
            }
        }
    }
}